fix: cilium webhook with templated values (#1437)
**What problem does this PR solve?**:
Its possible that users may want to use a templated values with their
custom Cilium ConfigMap to enable kube-proxy replacement feature.
This fixes the webhook to support that.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->